ui: storage edit: only LVM supports changing ext. snapshots for existing storages
Only allow changing external-snapshots for existing storages when type
is LVM, as for the directory based ones the option is marked as
`fixed` in the schema, which means it cannot be changed for existing
configuration entries.

While at it replace the defaultValue with checked, as the former is
not really used any way when using deleteEmpty.

Reported-by: Lukas Wagner <l.wagner@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>