qemu: network: adjust MTU emptyText to match new default behavior
Leaving the MTU field unset now defaults to the bridge MTU, rather
than 1500. Reflect this change by indicating the new behavior in the
emptyText of the MTU field. While we're at it, add a gettext call so
it can be translated. I've taken the same text as from the container
dialogue, so it should already use existing translations.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250717175012.606372-3-s.hanreich@proxmox.com