ui: disk storage selector: fix check for disabling format selector
The API returns an object where a format can map to 0 (unsupported) or
1 (supported). However, when deciding whether to disable the format
selector by counting the number of valid formats, the GUI only took
the number of entries into account, not the actual values, so it would
show the format selector even though there is only one entry with
value 1. Fix this by taking the values into account when counting the
number of valid formats.

Signed-off-by: Friedrich Weber <f.weber@proxmox.com>
Link: https://lore.proxmox.com/20250718102532.61149-1-f.weber@proxmox.com
 [TL: run through proxmox-biome format]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>