d/postinst: improve lvm filter mirgration
The old code worked but used a bit hard to read syntax in the sed
command that even confused my vim syntax highlighting into thinking
the quote is still open for the rest of the file.

Split into single and double quotes as needed, this also separates
backslash escaping for shell and for sed more clearly.

While at it ensure the new marker gets added, as otherwise one might
gain two entries.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>