configs: add udev helper for pinning virtual function names
This commmit adds a udev rule that triggers for every network device
that gets added. It checks if the network device is a VF and if the
parent device is pinned. If it is pinned, then generate a new name for
the VF which consists of the pinned name of the parent device, as well
as the index of the VF.

It relies on the network device driver exposing the information via
sysfs, which was the case in my tests for mlx5_core, igb and bnxt_en.

Specifically it checks if a device is a virtual function by checking
for the existence of:

  /sys/class/net/<iface>/device/physfn

It then follows that symlink and infers the vf index by looking at the
virtfnX symlinks in the folder above.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250722145223.351778-2-s.hanreich@proxmox.com