fix #6068: ui: utils: calculate and render host memory usage correctly
by adding the new memhost field, which is populated for VMs, and
using it if the guest is of type qemu and the field is numerical.

As a result, if the cluster is in a mixed PVE8 / PVE9 situation, for
example during a migration, we will not report any host memory usage, in
numbers or percent, as we don't get the memhost metric from the older
PVE8 hosts.

Fixes: #6068 (Node Search tab incorrect Host memory usage %)
Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>