make new Yew based UI opt-out for now
Just for initial testing, this way one can relatively easily switch
back to the old UI for comparison (of how bad it's state is).

While I really do not think this is strictly required, it allows
moving along a newer pve-manager without the Yew PVE GUI package, so
can be useful.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>