statd: re-cache IP link details every 15 minutes
While the commit adding this cache is right in that this data probably
won't change at all that often, adding a simple time based expiration
of the cache is not much extra code, and the normally 96 fork+exec of
ip link per day should really not be noticeable at all.

FWIW, if we wanted to make this more efficient and reduce the latency
on changes, we could hook into udev and touch some in-memory flag file
to signal pvestatd that it should requery this info. As this is an
internal mechanism, we can change it any time, so can still be done in
the future.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>