ui: ha: replace ha groups with ha node affinity rules
Introduce HA rules and replace the existing HA groups with the new HA
node affinity rules in the web interface.

The HA rules components are designed to be extendible for other new rule
types and allow users to display the errors of contradictory HA rules,
if there are any, in addition to the other basic CRUD operations.

HA rule ids are automatically generated with a 13 character UUID string
in the web interface, as also done for other concepts already, e.g.,
backup jobs, because coming up with future-proof rule ids that cannot be
changed later is not that user friendly. The HA rule's comment field is
meant to store that information instead.

Signed-off-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20250730175957.386674-30-d.kral@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>