http server: fix caching of i18n translation files, use mtime
By using the mtime of the parent directory where all locale i18n files
are located we ensure that every time a file is added/replaced/deleted
we will get a new value.

As both dpkg and install recreate existing files this should always
result in such an mtime update when either installing a new package or
directly installing the build artefacts to the root system
(circumventing dpkg).

While the explicit version would be slightly nicer, it's only relevant
for debugging and can be correlated with the local file and the one
from the debian package – which always has the time stamp from
the current d/changelog entry.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>