ui: rrd charts: drop info tool for now
Avoid rushing this, besides that it could be de-duplicated code-wise,
and might even go into widget-toolkit so that on use-site one only
passes a config option with the respective (get)text along, at least
if this should become a generic concept (which would be good to avoid
making it feel one-off "tacked-on).

It might also look better when being directly besides the title (on
the right side of that), not the legend.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>