buildsys: allow to opt-out strict online help reference checks
New pve-docs we now depend on supports a --allow-missing flag for the
asciidoc-pve scan-extjs command, which, if set, will silence errors
for missing onlineHelp references to warnings and fallback to a link
to the main single-page admin guide.

One can easily enable this by using `export ALLOW_MISSING=1` before
the build or bass that variable to make like:

 make ALLOW_MISSING=1 OnlineHelpInfo.js

in the www/manager6 source directory.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>