pve8to9: don't report already migrated files as needing to be migrated
the find command previously also found the already migrated rrd files
under `pve-{vm,node,storage}-9.0` and reported them as needing to
migrate them. the provided command to would of course not migrate them
so the warning persisted even after the command was run.

limit the find command to the old `pve2-` prefixed folders to prevent
that.

Reported-by: Friedrich Weber <f.weber@proxmoc.com>
Reported-by: Daniel Herzig <d.herzig@proxmox.com>
Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Tested-by: Daniel Herzig <d.herzig@proxmox.com>
Tested-by: Michael Köppl <m.koeppl@proxmox.com>
Link: https://lore.proxmox.com/20250801091336.68133-1-s.sterz@proxmox.com