pve8to9: check for legacy ipam db or mac cache files
Check if there are any not-yet-migrated ipam / mac cache files in
pmxcfs. Those should have been migrated over in the pve-network
postinst, but if something went wrong during this process we can
explicitly notify users here again to avoid any unpleasant suprises
after the upgrade.

If all nodes are on PVE 9, it is safe to delete the legacy files,
print a notice informing the users.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250801121029.201766-3-s.hanreich@proxmox.com