ha: affinity rules: use single data store and mirror it to child grids
This avoids loading the data twice, which is annoying especially on
high-latency links. Further, it ensures both views are always backed
by the exact same data, avoiding inconsistent state and thus e.g.
confusing warnings if something changed in the backend between on of
both loads.

To implement this add a new model that derives from the main one but
uses an in-memory proxy. Then move the "real" store out to the parent
component, where we need to manually initialise it as ExtJS panel are
more generic compared to grids–which always got a backing store.
Anyway, in the parent add a listener to copy any data to the in-memory
stores of the child grids for each affinity rule type. In the child
grid's relay any store load–e.g., after adding/changing/deleting a
rule–to the parent.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>