ui: node summary metrics: only show memory usage without ARC if ARC is more than 1 MiB
It looks rather odd to see the same size twice especially if one isn't
using ZFS in the first place, so use the simple heuristic of ARC
needing to be > 1 MiB to decide if we should render the ARC. Note that
the lowest possible minimum size for the ARC is 64 MiB, so this should
cover all real ZFS setups. FWIW, might also use a test that depends on
the order of magnitude difference between ARC size and memory used.

btw. this is also only theoretical true, with ZFS 2.3 the ARC is
registered for the kernel as reclaimable, so it might or might not
account to used, depending on the calculation and what the kernel
does, which changes frequently. If we get reports about this being odd
we should just remove it completely.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>