8 to 9 checks: warn if old udev-mac-pinning still present
We observed this in a host which has been updated through many major
versions (at least since PVE 4). The 70-persistent-net.rules was
used keeping the NICs with 'ethX' names. With the first reboot
into trixie the NICs got their predictable names, and networking was
broken (because the 'ethX' names are not present as altname the
support for this does not help in this case)

I could not reproduce the issue with a VM (there the
70-persistent-net.rules was still active and the NIC remained ethX),
so it might be a race during early boot.

In any case a warning makes sense here as it's becoming a very niche
combination, and thus is likely to cause more issues in the future.

Suggest to manually setup pinning as suggested in our documentation:
https://pve.proxmox.com/pve-docs/pve-admin-guide.html#network_override_device_names
seems sensible.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20250804183856.773378-1-s.ivanov@proxmox.com