d/tmpfiles: fix permission regression for /run/pve directory
There is a regression regarding the permission for the /run/pve
directory. In Proxmox VE 8, the directory had root:root 0755
permissions, being auto-created as the lxc-syscalld runtime directory.
In Proxmox VE 9, the permissions were restricted to root:root 0750,
but this leads to an issue with remote migration, when pveproxy tries
to access the mtunnel socket:

pveproxy[2484]: connect to 'unix/:/run/pve/ct-112.mtunnel' failed: Permission denied

Relax the permissions again by allowing the www-data group
read-access, so that pveproxy can access the socket.

This aligns the permissions with what /run/pve-cluster has.

Reported-by: Hannes Laimer <h.laimer@proxmox.com>
Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Tested-by: Hannes Laimer <h.laimer@proxmox.com>
Link: https://lore.proxmox.com/20250805100556.40874-1-f.ebner@proxmox.com