8 to 9 checks: check for removable grub-install
some upgrades result in unbootable systems, which can be traced back
to grub being installed in BOOTX64.efi, but not being upgraded by
grub-install. Refer the cases to the output of
`proxmox-boot-tool refresh` as it has a sensible check logic for those
cases. Some affected systems printed the warning of proxmox-boot-tool,
but it was lost in the large output of the dist-upgrade.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/all/20250808124540.1490294-3-s.ivanov@proxmox.com

FG: rename variable, set boot_ok correctly in last if

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>