8 to 9: support single-line repo entries with bracket options
I had a multi-arch setup here where I set an option to make apt check
only for amd64 on the PVE repo, i.e. :

 deb [arch=amd64] http://download.proxmox.com/debian/pve trixie pvetest

This was not detected by our regex and thus the pve8to9 did not
complain about the (for trixie) misspelled pvetest component.

Simply extend the regex to allow arbitrary string inside a bracket,
but do not add a match group for it, we would not do anything with it
for now anyway.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>