ui: storage edit: warn about disabling snapshot-as-volume-chain on LVM
Existing qcow2 volumes on a storage won't be handled correctly anymore
after the setting is turned off. The setting is already a fixed
storage setting for directory-based storages, so this is only relevant
for LVM. Could be improved by checking in the backend if there are any
qcow2 images and only allow turning it off if not, but this requires
changes to the on_update_hook() signature. Until then, warn in the
front-end.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20250812082824.30625-1-f.ebner@proxmox.com