fix #6599: ui: use font-awesome hdd icon instead of png
previously we used custom png rendered for storages in the add menu of
a pool, as well as for mountpoints of containers. these appeared
blurry on high resolution displays and since they were just the same
as the font-awesome icons anyway, use those directly. the ui already
loads font-awesome regardless, so there are no down-sides here.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20250826145836.302748-2-s.sterz@proxmox.com