ui: remove unused png icons and their asset files
a lot of icons are no longer used, so remove them and, in rare cases,
their accompanying css classes. all of these are png files and would,
thus, appear blurry if used. this should somewhat reduce the size of
the pve-manager package, but more importantly, discourage the use of
icons that would only appear blurry on modern hardware anyway.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20250826145836.302748-10-s.sterz@proxmox.com