API2Tools: rrd: remove O(n^2) lookup for keys
the idea was that we get any of the 'new' versions on lookup, but that
lead to iterating through possibly all keys. Since that was called for
each resource in e.g. /cluster/resources api call, the runtime was
O(n^2) for the number of resources.

To avoid that, simply look up the currently only valid key here which
makes this lookup much cheaper.

In my test setup with ~10000 guests, it reduces the time for a call
to /cluster/resources from ~22s to ~400ms

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Tested-by: Daniel Kral <d.kral@proxmox.com>
Link: https://lore.proxmox.com/20250905120627.2585826-2-d.csapak@proxmox.com