ui: resource tree: fix change detection
when trying to detect changes of resources, we compare a list of
properties of the existing nodes in the tree with the ones we got from
the api call, so that we can update only those that changed.

One of these properties is the 'text' one, which is calculated from e.g.
the vmid and name (or the name and host, depending on the type).

Sadly, when inserting/updating the node, we modified the text property
in every case, at least adding a '<span></span>' around the existing
text. This meant that every resource was updated every time instead of
only when something changed.

To fix this, remote the 'text' property from the to checked ones, and
add all the properties that are used to compile the text one.

This reduces the time of updateTree in my test-setup (~10000 guests)
when nothing changed from ~100ms to ~15ms and reduces scroll stutter
during such an update.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20250905120627.2585826-5-d.csapak@proxmox.com
(cherry picked from commit e19de81f5469eb2e4e45296ebeef16754abacc7a)
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>